move modern ChainListFooter & ChainListHeader
