+++ src/components/AssetSelectorModal/CollectibleList.js
@@ -25,8 +25,8 @@ import { chunk } from 'lodash';
 import { useTranslation } from 'translations/translate';
 
 // Components
-import ChainListHeader from 'components/modern/ChainListHeader';
-import ChainListFooter from 'components/modern/ChainListFooter';
+import ChainListHeader from 'components/lists/ChainListHeader';
+import ChainListFooter from 'components/lists/ChainListFooter';
 import EmptyStateParagraph from 'components/EmptyState/EmptyStateParagraph';
 
 // Selectors
