+++ src/screens/Assets/collectibles/CollectiblesTab.js
@@ -27,8 +27,8 @@ import { chunk } from 'lodash';
 import { useTranslation } from 'translations/translate';
 
 // Components
-import ChainListHeader from 'components/modern/ChainListHeader';
-import ChainListFooter from 'components/modern/ChainListFooter';
+import ChainListHeader from 'components/lists/ChainListHeader';
+import ChainListFooter from 'components/lists/ChainListFooter';
 import FloatingButtons from 'components/FloatingButtons';
 import Modal from 'components/Modal';
 import ReceiveModal from 'screens/Asset/ReceiveModal';
