+++ src/screens/Assets/investments/InvestmentsTab.js
@@ -28,8 +28,8 @@ import { useTranslationWithPrefix } from 'translations/translate';
 
 // Components
 import BalanceView from 'components/BalanceView';
-import ChainListHeader from 'components/modern/ChainListHeader';
-import ChainListFooter from 'components/modern/ChainListFooter';
+import ChainListHeader from 'components/lists/ChainListHeader';
+import ChainListFooter from 'components/lists/ChainListFooter';
 import FiatChangeView from 'components/modern/FiatChangeView';
 import FloatingButtons from 'components/FloatingButtons';
